refactoring

+++ app/src/main/java/pl/webfly/sensei/Finish.java
@@ -8,12 +8,6 @@ import android.widget.Button;
 import android.widget.TextView;
 import pl.webfly.sensei.trainer.Score;
 
-/**
- * Created by jbukowski on 2016-06-27
- * <p/>
- * Package: pl.webfly.sensei
- * Project: Tutorial
- */
 public class Finish extends AppCompatActivity implements View.OnClickListener {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -28,7 +22,7 @@ public class Finish extends AppCompatActivity implements View.OnClickListener {
 
 
         TextView scoreText = (TextView) findViewById(R.id.percentageScore);
-        scoreText.setText(Float.toString(score.getPercentageScore()) + "% (" + score.getCorrectAnswersQnt() + "/" + score.getAnsweredQuestionQnt() + ")");
+        scoreText.setText(Double.toString(score.getPercentageScore()) + "% (" + score.getCorrectAnswersQnt() + "/" + score.getAnsweredQuestionQnt() + ")");
     }
 
     @Override
